+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Upgraded `spotless-maven-plugin` to `3.1.0`.
 - Upgraded `org.eclipse.jdt/org.eclipse.jdt.core` to `3.44.0`.
 - Added `code-format` version `1.0.0` as the source for Java formatter file `java-code-style.xml`.
+- Added `Nullables.nonNullOrThrow` to throw the supplied `Throwable` instead of using `nonNullOrDefault`.
 
 ---